// src/app/models/membership.model.ts
export interface Membership{
  membershipType:string;
  membership: string;
  membershipPlan: string;
  firstName: string;
  lastName: string;
  speciality: string;
  password: string;
  confPassword: string;
  address: string;
  email: string;
  phone: string;
  profileUrl: string;
  studentDetails: string;
  addressLineOne:string;
  addressLineTwo:string;
  postalCode:string;
  state:string;
  city:string;
}